+++ DawnLib/src/API/Moons/MoonInfoBuilder.cs
@@ -1,4 +1,5 @@
 ﻿using System.Collections.Generic;
+using UnityEngine;
 
 namespace Dawn;
 public class MoonInfoBuilder : BaseInfoBuilder<DawnMoonInfo, SelectableLevel, MoonInfoBuilder>
@@ -36,9 +37,9 @@ public class MoonInfoBuilder : BaseInfoBuilder<DawnMoonInfo, SelectableLevel, Mo
         return this;
     }
 
-    public MoonInfoBuilder AddScene(NamespacedKey<IMoonSceneInfo> sceneKey, int weight, string assetBundlePath, string scenePath)
+    public MoonInfoBuilder AddScene(NamespacedKey<IMoonSceneInfo> sceneKey, AnimationClip shipLandingOverrideAnimation, AnimationClip shipTakeoffOverrideAnimation, int weight, string assetBundlePath, string scenePath)
     {
-        _scenes.Add(new CustomMoonSceneInfo(sceneKey, new SimpleProvider<int>(weight), assetBundlePath, scenePath));
+        _scenes.Add(new CustomMoonSceneInfo(sceneKey, shipLandingOverrideAnimation, shipTakeoffOverrideAnimation, new SimpleProvider<int>(weight), assetBundlePath, scenePath));
         return this;
     }
 
@@ -54,7 +55,6 @@ public class MoonInfoBuilder : BaseInfoBuilder<DawnMoonInfo, SelectableLevel, Mo
 
             returnString += charmander;
         }
-
         return returnString;
     }